requirements.txt
beautifulsoup4==4.3.2
Django==1.8
django-activeurl==0.1.8
django-appconf==1.0.1
django-classy-tags==0.6.1
django-debug-toolbar==1.3.0
django-extensions==1.5.5
django-haystack==2.3.1
django-model-utils==2.2
elasticsearch==1.4.0
lxml==3.4.2
MySQL-python==1.2.5
Pattern==2.6
Pillow==2.8.1
pygraphviz==1.2
python-memcached==1.54
python-ptrace==0.8.1
simplejson==3.6.5
six==1.9.0
sorl-thumbnail==12.2
sqlparse==0.1.14
urllib3==1.10.2
